Fixes a bug while swipping left-right on MainActivity.
When the user swipped left and right, the left-most or right-most
activity layout would be destroyed. When the page would be shown
again, the onCreateView method from each fragment was not using the
information from the ViewModel to configure either the loading status
display or the contents of the RecyclerView.